# src/openapi/api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: flaviolrita@hotmail.com
    name: "Flavio Lionel Rita "
  description: service of lambda orm
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: λORM Service
  version: 0.1.16
...
servers:
  - url: /
...
paths:
  /metadata:
    get:
      operationId: metadata
      parameters:
        - explode: true
          in: query
          name: query
          required: true
          schema:
            type: string
          style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Metadata"
          description: Ok
      security: []
      tags:
        - expression
      x-eov-operation-handler: controllers/ExpressionController
  /model:
    get:
      operationId: model
      parameters:
        - explode: true
          in: query
          name: query
          required: true
          schema:
            type: string
          style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/MetadataModel"
                type: array
          description: Ok
      security: []
      tags:
        - expression
      x-eov-operation-handler: controllers/ExpressionController
  /parameters:
    get:
      operationId: parameters
      parameters:
        - explode: true
          in: query
          name: query
          required: true
          schema:
            type: string
          style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/MetadataParameter"
                type: array
          description: Ok
      security: []
      tags:
        - expression
      x-eov-operation-handler: controllers/ExpressionController
  /constraints:
    get:
      operationId: constraints
      parameters:
        - explode: true
          in: query
          name: query
          required: true
          schema:
            type: string
          style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetadataConstraint"
          description: Ok
      security: []
      tags:
        - expression
      x-eov-operation-handler: controllers/ExpressionController
  /sentence:
    get:
      operationId: sentence
      parameters:
        - explode: true
          in: query
          name: query
          required: true
          schema:
            type: string
          style: form
        - explode: true
          in: query
          name: stage
          required: false
          schema:
            type: string
          style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetadataSentence"
          description: Ok
      security: []
      tags:
        - expression
      x-eov-operation-handler: controllers/ExpressionController
  /execute:
    post:
      operationId: execute
      parameters:
        - explode: true
          in: query
          name: query
          required: true
          schema:
            type: string
          style: form
        - explode: true
          in: query
          name: stage
          required: false
          schema:
            type: string
          style: form
      requestBody:
        content:
          application/json:
            schema:
              type: object
        required: false
      responses:
        "200":
          content:
            application/json:
              schema: {}
          description: Ok
      security: []
      tags:
        - expression
      x-eov-operation-handler: controllers/ExpressionController
...
components:
...
  schemas:
    Metadata:
      additionalProperties: false
      example:
        sentence: sentence
        number: 6.027456183070403
        children:
          - null
          - null
        name: name
        property: property
        alias: alias
        type: type
        fields:
          - readValue: readValue
            default: default
            nullable: true
            autoincrement: true
            name: name
            length: 0.8008281904610115
            type: type
            writeValue: writeValue
            enum: enum
            primaryKey: true
          - readValue: readValue
            default: default
            nullable: true
            autoincrement: true
            name: name
            length: 0.8008281904610115
            type: type
            writeValue: writeValue
            enum: enum
            primaryKey: true
        parameters:
          - name: name
            type: type
            value: ""
          - name: name
            type: type
            value: ""
        entity: entity
        relation:
          composite: true
          name: name
          from: from
          to: to
          targetComposite: true
          entity: entity
          weak: true
          target: target
      properties:
        name:
          type: string
        type:
          type: string
        entity:
          type: string
        children:
          items:
            $ref: "#/components/schemas/Metadata"
          type: array
        fields:
          items:
            $ref: "#/components/schemas/Property"
          type: array
        property:
          type: string
        parameters:
          items:
            $ref: "#/components/schemas/Parameter"
          type: array
        relation:
          $ref: "#/components/schemas/Relation"
        sentence:
          type: string
        alias:
          type: string
        number:
          format: double
          type: number
      required:
        - name
        - type
      type: object
    Property:
      additionalProperties: false
      example:
        readValue: readValue
        default: default
        nullable: true
        autoincrement: true
        name: name
        length: 0.8008281904610115
        type: type
        writeValue: writeValue
        enum: enum
        primaryKey: true
      properties:
        name:
          type: string
        type:
          type: string
        length:
          format: double
          type: number
        nullable:
          type: boolean
        primaryKey:
          type: boolean
        autoincrement:
          type: boolean
        default:
          type: string
        readValue:
          type: string
        writeValue:
          type: string
        enum:
          type: string
      required:
        - name
        - type
      type: object
    Parameter:
      additionalProperties: false
      example:
        name: name
        type: type
        value: ""
      properties:
        name:
          type: string
        type:
          type: string
        value: {}
      required:
        - name
        - type
      type: object
    RelationType:
      enum:
        - oneToMany
        - manyToOne
        - oneToOne
      type: string
    Relation:
      additionalProperties: false
      example:
        composite: true
        name: name
        from: from
        to: to
        targetComposite: true
        entity: entity
        weak: true
        target: target
      properties:
        name:
          type: string
        type:
          $ref: "#/components/schemas/RelationType"
        composite:
          type: boolean
        from:
          type: string
        entity:
          type: string
        to:
          type: string
        weak:
          type: boolean
        target:
          type: string
        targetComposite:
          type: boolean
      required:
        - entity
        - from
        - name
        - to
        - type
      type: object
    MetadataModel:
      additionalProperties: false
      example:
        name: name
        type: type
        childs:
          - null
          - null
      properties:
        name:
          type: string
        type:
          type: string
        childs:
          items:
            $ref: "#/components/schemas/MetadataModel"
          type: array
      required:
        - name
        - type
      type: object
    MetadataParameter:
      additionalProperties: false
      example:
        name: name
        type: type
        childs:
          - null
          - null
      properties:
        name:
          type: string
        type:
          type: string
        childs:
          items:
            $ref: "#/components/schemas/MetadataParameter"
          type: array
      required:
        - name
        - type
      type: object
    Constraint:
      additionalProperties: false
      example:
        condition: condition
        message: message
      properties:
        message:
          type: string
        condition:
          type: string
      required:
        - condition
        - message
      type: object
    MetadataConstraint:
      additionalProperties: false
      example:
        constraints:
          - condition: condition
            message: message
          - condition: condition
            message: message
        childs:
          - null
          - null
        entity: entity
      properties:
        entity:
          type: string
        constraints:
          items:
            $ref: "#/components/schemas/Constraint"
          type: array
        childs:
          items:
            $ref: "#/components/schemas/MetadataConstraint"
          type: array
      required:
        - constraints
        - entity
      type: object
    MetadataSentence:
      additionalProperties: false
      example:
        sentence: sentence
        dialect: dialect
        dataSource: dataSource
        childs:
          - null
          - null
        entity: entity
      properties:
        entity:
          type: string
        dialect:
          type: string
        dataSource:
          type: string
        sentence:
          type: string
        childs:
          items:
            $ref: "#/components/schemas/MetadataSentence"
          type: array
      required:
        - dataSource
        - dialect
        - entity
        - sentence
      type: object
...
  securitySchemes: {}